requirements.txt
Django==2.2
django-rest-authtoken==2.1.3
djangorestframework==3.12.1
protobuf==3.13.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1

